lua/utils: `textutils` module for text formatting
This patch adds a `textutils` module, which provides utility functions
for manipulating and formatting text.

The `dedent` function removes common leading whitespace from multiline
strings, similar to Python's `textwrap.dedent()`, and is useful for
handling string literals ([[...]]).

The `toline` function  condenses multiline strings into a single line by
removing newlines and extra spaces, making it easier to format error
messages or other long text sections.

Needed for #10636

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=internal
